# backend/internal/data/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package db

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type PackingList struct {
	ID          uuid.UUID
	CreatedAt   time.Time
	UpdatedAt   time.Time
	UserID      uuid.UUID
	Name        string
	Description string
	Status      string
	DueDate     pgtype.Date
	Tags        []string
	Days        int32
}

type PackingListItem struct {
	ID            uuid.UUID
	CreatedAt     time.Time
	UpdatedAt     time.Time
	PackingListID uuid.UUID
	Name          string
	Category      string
	Quantity      int32
	IsPacked      bool
	Notes         string
}

type User struct {
	ID                    uuid.UUID
	CreatedAt             time.Time
	UpdatedAt             time.Time
	Username              string
	Email                 string
	PasswordHash          string
	IsAdmin               bool
	StripeCustomerID      *string
	StripeSubscriptionID  *string
	SubscriptionStartDate pgtype.Timestamp
	SubscriptionEndedDate pgtype.Timestamp
}

type UserActionToken struct {
	ID        uuid.UUID
	CreatedAt time.Time
	ExpiresAt time.Time
	UserID    uuid.UUID
	Token     []byte
	Action    string
}

type UserSession struct {
	ID        uuid.UUID
	CreatedAt time.Time
	ExpiresAt time.Time
	UserID    uuid.UUID
	Token     []byte
}