# backend/internal/web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a standard Rest API template",
        "title": "hookfeed",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/packing-lists": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List all PackingLists",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackingList"
                ],
                "summary": "List all PackingLists",
                "parameters": [
                    {
                        "enum": [
                            "created_at"
                        ],
                        "type": "string",
                        "description": "order by",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "The number of items to skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "The number of items to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of PackingLists",
                        "schema": {
                            "$ref": "#/definitions/dtos.PaginationResponse-dtos_PackingListSummary"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new PackingList",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackingList"
                ],
                "summary": "Create a new PackingList",
                "parameters": [
                    {
                        "description": "The PackingList details",
                        "name": "PackingList",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.PackingListCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.PackingList"
                        }
                    }
                }
            }
        },
        "/v1/packing-lists/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a PackingList",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackingList"
                ],
                "summary": "Get a PackingList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The PackingList ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.PackingList"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a PackingList",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackingList"
                ],
                "summary": "Update a PackingList",
                "parameters": [
                    {
                        "description": "The PackingList details",
                        "name": "PackingList",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.PackingListUpdate"
                        }
                    },
                    {
                        "type": "string",
                        "description": "The PackingList ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.PackingList"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a PackingList",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackingList"
                ],
                "summary": "Delete a PackingList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The PackingList ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/packing-lists/{id}/items": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new PackingListItem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackingListItem"
                ],
                "summary": "Create a new PackingListItem",
                "parameters": [
                    {
                        "description": "The PackingListItem details",
                        "name": "PackingListItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.PackingListItemCreate"
                        }
                    },
                    {
                        "type": "string",
                        "description": "The PackingList ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.PackingListItem"
                        }
                    }
                }
            }
        },
        "/v1/packing-lists/{id}/items/{item-id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a PackingListItem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackingListItem"
                ],
                "summary": "Update a PackingListItem",
                "parameters": [
                    {
                        "description": "The PackingListItem details",
                        "name": "PackingListItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.PackingListItemUpdate"
                        }
                    },
                    {
                        "type": "string",
                        "description": "The PackingList ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The PackingListItem ID",
                        "name": "item-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.PackingListItem"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a PackingListItem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PackingListItem"
                ],
                "summary": "Delete a PackingListItem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The PackingList ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The PackingListItem ID",
                        "name": "item-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.PackingList": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "days": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.PackingListItem"
                    }
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dtos.PackingListCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "days": {
                    "type": "integer",
                    "x-nullable": true
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string",
                    "x-nullable": true
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "x-nullable": true
                }
            }
        },
        "dtos.PackingListItem": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isPacked": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "dtos.PackingListItemCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "dtos.PackingListItemUpdate": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "x-nullable": true
                },
                "isPacked": {
                    "type": "boolean",
                    "x-nullable": true
                },
                "name": {
                    "type": "string",
                    "x-nullable": true
                },
                "notes": {
                    "type": "string",
                    "x-nullable": true
                },
                "quantity": {
                    "type": "integer",
                    "x-nullable": true
                }
            }
        },
        "dtos.PackingListSummary": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "days": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "itemCount": {
                    "type": "integer"
                },
                "itemPackedCount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dtos.PackingListUpdate": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "integer",
                    "x-nullable": true
                },
                "description": {
                    "type": "string",
                    "x-nullable": true
                },
                "dueDate": {
                    "type": "string",
                    "x-nullable": true
                },
                "name": {
                    "type": "string",
                    "x-nullable": true
                },
                "status": {
                    "type": "string",
                    "x-nullable": true
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "x-nullable": true
                }
            }
        },
        "dtos.PaginationResponse-dtos_PackingListSummary": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.PackingListSummary"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}